last linting most W and C are fixed only E left for import error
